[+] добавлена поддержка VKApp 4.6 [+] добавлено изменение сафари в ночной теме

+++ ColoredVK/Headers/VKViews.h
@@ -6,7 +6,7 @@
 //
 
 @interface BackgroundView : UILabel
-@property (assign, nonatomic) int cornerRadius;
+@property (assign, nonatomic) NSInteger cornerRadius;
 @end
 
 @interface VKMImageButton : UIButton
